TargetSearchData refactoring
This package only contains examplary data files for TargetSearch, and as
such does not provide any function, code examples, data structures, &c.
For historical reasons, this package used to depend on TargetSearch
which created an unnecessary chicken-egg situation during installation.

This commit removes the TargetSearch dependency so this package can be
installed (and R-checked) independently of TargetSearch.

The objects/data structures which used to be in this package have been moved
to TargetSearch.